0-putchar.c

+++ functions_nested_loops/0-putchar.c
@@ -10,7 +10,7 @@ int main(void)
 	char *str = "_putchar\n";
 	int i;
 
-	for (i = 0; i != strlen(str); i++)
+	for (i = 0; i != (int)strlen(str); i++)
 		_putchar(str[i]);
 	return (0);
 }
